# swaggerDesign.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger Book Tracker",
        "description": "This is a book tracker server.\n",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "angelahurd@byui.edu"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "schemes": ["http"],
    "tags": [{
            "name": "book",
            "description": "Book Collection"
        },
        {
            "name": "user",
            "description": "Operations about user"
        }
    ],
    "paths": {
        "/book": {
            "get": {
                "tags": [
                    "book"
                ],
                "summary": "Get all the books",
                "operationId": "getAllBooks",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Books"
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "$ref": "#/components/schemas/Books"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Book not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "book"
                ],
                "summary": "Add a new book",
                "operationId": "createBook",
                "requestBody": {
                    "$ref": "#/components/requestBodies/Books"
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [{
                    "bookTracker_auth": [
                        "write:books",
                        "read:books"
                    ]
                }]
            }
        },
        "/book/{_id}": {
            "get": {
                "tags": [
                    "book"
                ],
                "summary": "Find book by book ID",
                "description": "Returns a single book",
                "operationId": "getBook",
                "parameters": [{
                    "name": "_id",
                    "in": "path",
                    "description": "Name the book needed",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Books"
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "$ref": "#/components/schemas/Books"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Book not found"
                    }
                },
                "security": [{
                    "api_key": []
                }]
            },
            "put": {
                "tags": [
                    "book"
                ],
                "summary": "Update an existing book",
                "operationId": "updateBook",
                "parameters": [{
                    "name": "_id",
                    "in": "path",
                    "description": "Book name to update",
                    "required": true,
                    "style": "simple",
                    "explode": false,
                    "schema": {
                        "type": "string"
                    }
                }],
                "requestBody": {
                    "$ref": "#/components/requestBodies/Books"
                },
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Book not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                },
                "security": [{
                    "bookTracker_auth": [
                        "write:books",
                        "read:books"
                    ]
                }]
            },
            "delete": {
                "tags": [
                    "book"
                ],
                "summary": "Delete a book",
                "operationId": "deleteBook",
                "parameters": [{
                    "name": "_id",
                    "in": "path",
                    "description": "Book name to delete",
                    "required": true,
                    "style": "simple",
                    "explode": false,
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Book not found"
                    }
                },
                "security": [{
                    "bookTracker_auth": [
                        "write:books",
                        "read:books"
                    ]
                }]
            }
        },
        "/user": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Create user",
                "description": "This can only be done by the logged in user.",
                "operationId": "createUser",
                "requestBody": {
                    "description": "Created user object",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/user/login": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Logs user into the system",
                "operationId": "loginUser",
                "parameters": [{
                        "name": "username",
                        "in": "query",
                        "description": "The user name for login",
                        "required": true,
                        "style": "form",
                        "explode": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "The password for login in clear text",
                        "required": true,
                        "style": "form",
                        "explode": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "headers": {
                            "X-Rate-Limit": {
                                "description": "calls per hour allowed by the user",
                                "style": "simple",
                                "explode": false,
                                "schema": {
                                    "type": "integer",
                                    "format": "int32"
                                }
                            },
                            "X-Expires-After": {
                                "description": "date in UTC when token expires",
                                "style": "simple",
                                "explode": false,
                                "schema": {
                                    "type": "string",
                                    "format": "date-time"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid username/password supplied"
                    }
                }
            }
        },
        "/user/logout": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Logs out current logged in user session",
                "operationId": "logoutUser",
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/user/{username}": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Get user by user name",
                "operationId": "getUserByName",
                "parameters": [{
                    "name": "username",
                    "in": "path",
                    "description": "The name that needs to be fetched. Use user1 for testing.",
                    "required": true,
                    "style": "simple",
                    "explode": false,
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid username supplied"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "user"
                ],
                "summary": "Update user",
                "description": "This can only be done by the logged in user.",
                "operationId": "updateUser",
                "parameters": [{
                    "name": "username",
                    "in": "path",
                    "description": "name that need to be updated",
                    "required": true,
                    "style": "simple",
                    "explode": false,
                    "schema": {
                        "type": "string"
                    }
                }],
                "requestBody": {
                    "description": "Updated user object",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "400": {
                        "description": "Invalid user supplied"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "user"
                ],
                "summary": "Delete user",
                "description": "This can only be done by the logged in user.",
                "operationId": "deleteUser",
                "parameters": [{
                    "name": "username",
                    "in": "path",
                    "description": "The name that needs to be deleted",
                    "required": true,
                    "style": "simple",
                    "explode": false,
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "400": {
                        "description": "Invalid username supplied"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Books": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "integer",
                        "format": "string"
                    },
                    "bookName": {
                        "type": "string",
                        "example": "Where the Sidewalk Ends"
                    },
                    "bookAuthor": {
                        "example": "Shel Silverstein"
                    },
                    "numberOfPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "bookDescription": {
                        "type": "string",
                        "example": "A masterful collection of poems and drawings that stretch the bounds of imagination."
                    },
                    "bookCoverImg": {
                        "type": "string",
                        "example": "https://imgur.com/h9zfeo1"
                    }
                },
                "xml": {
                    "name": "Book"
                }
            },
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "username": {
                        "type": "string"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "booksRead": {
                        "type": "string"
                    }
                },
                "xml": {
                    "name": "User"
                }
            }
        },
        "requestBodies": {
            "Books": {
                "description": "Books object that needs to be added to the user",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Books"
                        }
                    },
                    "application/xml": {
                        "schema": {
                            "$ref": "#/components/schemas/Books"
                        }
                    }
                },
                "required": true
            },
            "UserArray": {
                "description": "List of user object",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "required": false
            }
        },
        "securitySchemes": {
            "bookTracker_auth": {
                "type": "oauth2",
                "flows": {
                    "implicit": {
                        "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
                        "scopes": {
                            "write:books": "modify books in your account",
                            "read:books": "read your books"
                        }
                    }
                }
            },
            "api_key": {
                "type": "apiKey",
                "name": "api_key",
                "in": "header"
            }
        }
    }
}